refactor/#157: 세그먼트 뷰 한 번만 init
