+++ Archive/ViewControllers/MusicLoadingVC/MusicSegmentViewController.swift
@@ -1,0 +1,12 @@
+//
+//  MusicSegmentViewController.swift
+//  Archive
+//
+//  Created by 이수현 on 2/19/25.
+//
+
+import UIKit
+
+class MusicSegmentViewController: UIViewController {
+    
+}
