+++ Archive/Views/Music_LoadingViews/MusicSegmentView.swift
@@ -8,11 +8,18 @@ class MusicSegmentView: UIView {
         let items = ["다음 트랙", "가사", "추천 콘텐츠"]
         let segmentedControl = UISegmentedControl(items: items)
         segmentedControl.selectedSegmentIndex = 0
-        segmentedControl.backgroundColor = .black
+        segmentedControl.backgroundColor = .clear
         segmentedControl.selectedSegmentTintColor = .clear
         
-        segmentedControl.setTitleTextAttributes([.foregroundColor: UIColor.white, .font: UIFont.boldSystemFont(ofSize: 16)], for: .selected)
-        segmentedControl.setTitleTextAttributes([.foregroundColor: UIColor.gray, .font: UIFont.systemFont(ofSize: 16)], for: .normal)
+        segmentedControl.setTitleTextAttributes([
+            .foregroundColor: UIColor.white,
+            .font: UIFont.customFont(font: .SFPro, ofSize: 14, rawValue: 400),
+            .baselineOffset: 7], for: .selected)
+        
+        segmentedControl.setTitleTextAttributes([
+            .foregroundColor: UIColor.white_70,
+            .font: UIFont.customFont(font: .SFPro, ofSize: 14, rawValue: 400),
+            .baselineOffset: 7], for: .normal)
 
         segmentedControl.setBackgroundImage(UIImage(), for: .normal, barMetrics: .default)
         segmentedControl.setDividerImage(UIImage(), forLeftSegmentState: .normal, rightSegmentState: .normal, barMetrics: .default)
@@ -22,10 +29,10 @@ class MusicSegmentView: UIView {
     let nextTrackCollectionView: UICollectionView = {
         let layout = UICollectionViewFlowLayout()
         layout.scrollDirection = .vertical
-        layout.itemSize = CGSize(width: UIScreen.main.bounds.width, height: 60)
-        layout.minimumLineSpacing = 10
+        layout.itemSize = CGSize(width: UIScreen.main.bounds.width, height: 50)
+        layout.minimumLineSpacing = 12
         let collectionView = UICollectionView(frame: .zero, collectionViewLayout: layout)
-        collectionView.backgroundColor = .black
+        collectionView.backgroundColor = .clear
         return collectionView
     }()
 
@@ -34,7 +41,7 @@ class MusicSegmentView: UIView {
         layout.scrollDirection = .vertical
         layout.itemSize = CGSize(width: UIScreen.main.bounds.width, height: 40)
         let collectionView = UICollectionView(frame: .zero, collectionViewLayout: layout)
-        collectionView.backgroundColor = .black
+        collectionView.backgroundColor = .clear
         return collectionView
     }()
     private lazy var normalUnderbar = UIView().then{
@@ -54,7 +61,7 @@ class MusicSegmentView: UIView {
         layout.scrollDirection = .vertical
         layout.itemSize = CGSize(width: UIScreen.main.bounds.width, height: 60)
         let collectionView = UICollectionView(frame: .zero, collectionViewLayout: layout)
-        collectionView.backgroundColor = .black
+        collectionView.backgroundColor = .clear
         return collectionView
     }()
 
@@ -64,7 +71,7 @@ class MusicSegmentView: UIView {
         layout.itemSize = CGSize(width: 140, height: 186)
         layout.minimumLineSpacing = 10
         let collectionView = UICollectionView(frame: .zero, collectionViewLayout: layout)
-        collectionView.backgroundColor = .black
+        collectionView.backgroundColor = .clear
         return collectionView
     }()
 
@@ -86,7 +93,7 @@ class MusicSegmentView: UIView {
 
     override init(frame: CGRect) {
         super.init(frame: frame)
-        backgroundColor = .black
+        backgroundColor = .black_100
         setupViews()
         setupConstraints()
     }
